refactor: relocate sendRelayMessage function to relayMessage module and update imports across related files for improved organization
