+++ core/mpc/keysign/execute/index.ts
@@ -16,7 +16,7 @@ import { Keyshare } from '../../lib/keyshare'
 import { SignSession } from '../../lib/signSession'
 import { deleteRelayMessage } from '../../relayMessage/delete'
 import { getRelayMessages } from '../../relayMessage/get'
-import { sendRelayMessage } from '../../sendRelayMessage'
+import { sendRelayMessage } from '../../relayMessage/send'
 import { sleep } from '../../sleep'
 import { uploadSetupMessage } from '../../uploadSetupMessage'
 import { KeysignSignature } from '../KeysignSignature'
